requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
